notify-server/requirements/dev.txt
-r base.txt
anyio==3.6.2
autoflake==2.1.1
black==23.3.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ciscoconfparse==1.7.24
cryptography==41.0.1
click==8.1.3
deprecat==2.1.1
dnspython==2.3.0
fortigate-api==1.2.4
future==0.18.3
fastapi==0.95.1
flake8==6.0.0
h11==0.14.0
idna==3.4
loguru==0.6.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
netmiko==4.2.0
ntc-templates==3.4.0
paramiko==3.2.0
passlib==1.7.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
importlib-metadata==6.5.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
anyio==3.6.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.1
sniffio==1.3.0
typing_extensions==4.5.0
coverage==7.2.3
pytest-cov>=3.0.0
typing==3.7.4.3
wrapt==1.15.0
python-telegram-bot==20.4
telegram==0.0.1
splunk-sdk==1.7.3
pymsteams==0.2.2
